backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
sqlalchemy==2.0.32
alembic==1.13.2
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
pillow==10.4.0
deepface==0.0.93
numpy==1.26.4
python-dotenv==1.0.1
cloudinary==1.41.0
websockets==12.0
pydantic[email]==2.8.2
pydantic-settings==2.4.0
scikit-learn==1.5.1
httpx==0.27.0
